bring adler32 to latest code changes
 lib/adler32.c
	fix typo
 lib/alpha/adler32.c
	comments, typos, renames, 64 bit is now generic
 lib/arm/adler32.c
	disaable BE NEON, add iwmmxt, disable ARM DSP, 
 lib/generic/adler32.c
	make NO_DIVIDE reduce, add 64 bit generic code, add gap filler 16 to VMIN_WORK 
 lib/ia64/adler32.c
	totally overhaul
 lib/mips/adler32.c
	fix typos, endianess, overhaul
 lib/ppc/adler32.c
	brush up
 lib/sparc/adler32.c
	totally overhaul
 lib/x86/adler32.c
	little touch ups

 lib/ia64/ia64.h
	add instruction
 lib/sparc/sparc_vis.h
	add helper
 lib/ppc/ppc_altivec.h
	add func to load a single u32
 lib/generic/strnlen.c
	fix typo
 lib/ppc/cpy_rest.c
	fix compile error
 lib/udpfromto.c
	remove warning

+++ lib/udpfromto.c
@@ -429,7 +429,7 @@ ssize_t recvmfromto_pre(int s, struct mfromto *info, size_t len, int flags)
 	}
 #endif
 
-	return err >= 0 ? i : -((ssize_t)(i + 1));
+	return err >= 0 ? (ssize_t)i : -((ssize_t)(i + 1));
 }
 
 void recvmfromto_post(struct mfromto *info, size_t len)
